requirements.txt
fastapi~=0.110.1
uvicorn[standard]~=0.29.0
python-dotenv~=1.0.1
sqlmodel~=0.0.16
pyjwt~=2.8.0
passlib~=1.7.4
python-multipart
alembic~=1.13.1
bcrypt~=4.2.1
pydantic~=2.5.3
SQLAlchemy~=2.0.29
psycopg2-binary~=2.9.10
pillow~=11.0.0
